@@ -66,8 +66,8 @@ dependencies {
 }
 
 compileJava {
-    sourceCompatibility = 17
-    targetCompatibility = 17
+    sourceCompatibility = 21
+    targetCompatibility = 21
 }
 
 tasks.withType(JavaCompile) {
