@@ -4,12 +4,12 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
-import javax.annotation.Resource;
-
-import org.allGraphQLCases.server.STP_AllFieldCasesWithIdSubtype_STS;
 import org.allGraphQLCases.server.DataFetchersDelegateAllFieldCasesWithIdSubtype;
+import org.allGraphQLCases.server.STP_AllFieldCasesWithIdSubtype_STS;
 import org.dataloader.BatchLoaderEnvironment;
 import org.springframework.stereotype.Component;
+
+import jakarta.annotation.Resource;
 
 @Component
 public class DataFetchersDelegateAllFieldCasesWithIdSubtypeImpl
@@ -32,17 +32,19 @@ public class DataFetchersDelegateAllFieldCasesWithIdSubtypeImpl
 		List<STP_AllFieldCasesWithIdSubtype_STS> list = new ArrayList<>(keys.size());
 
 		for (UUID id : keys) {
-			STP_AllFieldCasesWithIdSubtype_STS item = generator.generateInstance(STP_AllFieldCasesWithIdSubtype_STS.class);
+			STP_AllFieldCasesWithIdSubtype_STS item = generator
+					.generateInstance(STP_AllFieldCasesWithIdSubtype_STS.class);
 			item.setId(id);
 
 			Object context = environment.getKeyContexts().get(id);
 			if (context != null) {
 
-				if (!(context instanceof KeyContext))
+				if (!(context instanceof KeyContext)) {
 					throw new RuntimeException(
 							"The DataFetchersDelegateAllFieldCasesWithIdSubtypeImpl's batchLoader received a bad context: it's an instance of "
 									+ context.getClass().getName() + ", but the expected class is "
 									+ KeyContext.class.getName());
+				}
 
 				KeyContext keyContext = (KeyContext) context;
 
