@@ -6,8 +6,6 @@ package org.allGraphQLCases.server.impl;
 import java.util.List;
 import java.util.UUID;
 
-import javax.annotation.Resource;
-
 import org.allGraphQLCases.server.DataFetchersDelegateCharacter;
 import org.allGraphQLCases.server.SEP_Episode_SES;
 import org.allGraphQLCases.server.SIP_Character_SIS;
@@ -15,6 +13,7 @@ import org.dataloader.BatchLoaderEnvironment;
 import org.springframework.stereotype.Component;
 
 import graphql.schema.DataFetchingEnvironment;
+import jakarta.annotation.Resource;
 
 /**
  * @author etienne-sf
@@ -28,17 +27,17 @@ public class DataFetchersDelegateCharacterImpl implements DataFetchersDelegateCh
 
 	@Override
 	public List<SIP_Character_SIS> friends(DataFetchingEnvironment dataFetchingEnvironment, SIP_Character_SIS source) {
-		return this.generator.generateInstanceList(SIP_Character_SIS.class, 4);
+		return generator.generateInstanceList(SIP_Character_SIS.class, 4);
 	}
 
 	@Override
 	public List<SEP_Episode_SES> appearsIn(DataFetchingEnvironment dataFetchingEnvironment, SIP_Character_SIS source) {
-		return this.generator.generateInstanceList(SEP_Episode_SES.class, 2);
+		return generator.generateInstanceList(SEP_Episode_SES.class, 2);
 	}
 
 	@Override
 	public List<SIP_Character_SIS> batchLoader(List<UUID> keys, BatchLoaderEnvironment environment) {
-		return this.generator.generateInstanceList(SIP_Character_SIS.class, keys.size());
+		return generator.generateInstanceList(SIP_Character_SIS.class, keys.size());
 	}
 
 }
