@@ -3,8 +3,6 @@ package org.allGraphQLCases.server.impl;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
 
-import javax.annotation.Resource;
-
 import org.allGraphQLCases.server.DataFetchersDelegateReservedJavaKeywordAllFieldCases;
 import org.allGraphQLCases.server.SIP_WithID_SIS;
 import org.allGraphQLCases.server.STP_Human_STS;
@@ -14,6 +12,7 @@ import org.dataloader.DataLoader;
 import org.springframework.stereotype.Component;
 
 import graphql.schema.DataFetchingEnvironment;
+import jakarta.annotation.Resource;
 
 @Component
 public class DataFetchersDelegateReservedJavaKeywordAllFieldCasesImpl
@@ -25,25 +24,25 @@ public class DataFetchersDelegateReservedJavaKeywordAllFieldCasesImpl
 	@Override
 	public CompletableFuture<SIP_WithID_SIS> _implements(DataFetchingEnvironment dataFetchingEnvironment,
 			DataLoader<UUID, SIP_WithID_SIS> dataLoader, STP_ReservedJavaKeywordAllFieldCases_STS origin) {
-		return CompletableFuture.completedFuture(this.generator.generateInstance(SIP_WithID_SIS.class));
+		return CompletableFuture.completedFuture(generator.generateInstance(SIP_WithID_SIS.class));
 	}
 
 	@Override
 	public CompletableFuture<STP_Human_STS> _int(DataFetchingEnvironment dataFetchingEnvironment,
 			DataLoader<UUID, STP_Human_STS> dataLoader, STP_ReservedJavaKeywordAllFieldCases_STS origin) {
-		return CompletableFuture.completedFuture(this.generator.generateInstance(STP_Human_STS.class));
+		return CompletableFuture.completedFuture(generator.generateInstance(STP_Human_STS.class));
 	}
 
 	@Override
 	public SUP_AnyCharacter_SUS _interface(DataFetchingEnvironment dataFetchingEnvironment,
 			STP_ReservedJavaKeywordAllFieldCases_STS origin) {
-		return this.generator.generateInstance(SUP_AnyCharacter_SUS.class);
+		return generator.generateInstance(SUP_AnyCharacter_SUS.class);
 	}
 
 	@Override
 	public CompletableFuture<SIP_WithID_SIS> nonJavaKeywordField(DataFetchingEnvironment dataFetchingEnvironment,
 			DataLoader<UUID, SIP_WithID_SIS> dataLoader, STP_ReservedJavaKeywordAllFieldCases_STS origin) {
-		return CompletableFuture.completedFuture(this.generator.generateInstance(SIP_WithID_SIS.class));
+		return CompletableFuture.completedFuture(generator.generateInstance(SIP_WithID_SIS.class));
 	}
 
 }
