@@ -6,12 +6,12 @@ package org.allGraphQLCases.server.impl;
 import java.util.List;
 import java.util.UUID;
 
-import javax.annotation.Resource;
-
 import org.allGraphQLCases.server.DataFetchersDelegateWithID;
 import org.allGraphQLCases.server.SIP_WithID_SIS;
 import org.dataloader.BatchLoaderEnvironment;
 import org.springframework.stereotype.Component;
+
+import jakarta.annotation.Resource;
 
 /**
  * @author etienne-sf
